Refactor Button and Modal components
Restructure the code so that these components are reusable.

Also use the new Modal component for QR code modal.
